backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.8.0
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
httpx==0.27.2
traclus-python==1.0.0
